service: mii: separate mii types into their own file
