+++ src/core/hle/service/mii/mii_manager.h
@@ -6,7 +6,10 @@
 #include <vector>
 
 #include "core/hle/result.h"
-#include "core/hle/service/mii/types.h"
+#include "core/hle/service/mii/mii_types.h"
+#include "core/hle/service/mii/types/char_info.h"
+#include "core/hle/service/mii/types/store_data.h"
+#include "core/hle/service/mii/types/ver3_store_data.h"
 
 namespace Service::Mii {
 
@@ -25,7 +28,7 @@ public:
     CharInfo BuildDefault(std::size_t index);
     CharInfo ConvertV3ToCharInfo(const Ver3StoreData& mii_v3) const;
     bool ValidateV3Info(const Ver3StoreData& mii_v3) const;
-    std::vector<MiiInfoElement> GetDefault(SourceFlag source_flag);
+    std::vector<CharInfoElement> GetDefault(SourceFlag source_flag);
     Result GetIndex(const CharInfo& info, u32& index);
 
     // This is nn::mii::detail::Ver::StoreDataRaw::BuildFromStoreData
@@ -33,6 +36,15 @@ public:
 
     // This is nn::mii::detail::NfpStoreDataExtentionRaw::SetFromStoreData
     NfpStoreDataExtension SetFromStoreData(const CharInfo& mii) const;
+
+    struct MiiDatabase {
+        u32 magic{}; // 'NFDB'
+        std::array<StoreData, 0x64> miis{};
+        INSERT_PADDING_BYTES(1);
+        u8 count{};
+        u16 crc{};
+    };
+    static_assert(sizeof(MiiDatabase) == 0x1A98, "MiiDatabase has incorrect size.");
 
 private:
     const Common::UUID user_id{};
