+++ src/core/hle/service/mii/types/raw_data.h
@@ -5,7 +5,7 @@
 
 #include <array>
 
-#include "core/hle/service/mii/types.h"
+#include "core/hle/service/mii/mii_types.h"
 
 namespace Service::Mii::RawData {
