+++ src/core/hle/service/mii/types/store_data.cpp
@@ -1,0 +1,6 @@
+// SPDX-FileCopyrightText: Copyright 2023 yuzu Emulator Project
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+#include "core/hle/service/mii/types/store_data.h"
+
+namespace Service::Mii {} // namespace Service::Mii
